feat: rename DoubleDatePicker component
