+++ packages/components/src/components/index.ts
@@ -4,11 +4,12 @@ export { default as Breadcrumb } from './atoms/breadcrumb/Breadcrumb'
 export { default as Breadcrumbs } from './atoms/breadcrumb/Breadcrumbs'
 export { default as Chip } from './atoms/Chip/Chip'
 export { default as CloseComponent } from './atoms/closeComponent/CloseComponent'
-export { default as Checkbox } from './atoms/form/CheckboxWithoutError'
-export { default as DatePicker } from './atoms/form/DatePickerWithoutError'
-export { default as DatePickerError } from './atoms/form/DatePicker'
+export { default as CheckboxWithoutError } from './atoms/form/CheckboxWithoutError'
+export { default as Checkbox } from './atoms/form/Checkbox'
+export { default as DatePickerWithoutError } from './atoms/form/DatePickerWithoutError'
+export { default as DatePicker } from './atoms/form/DatePicker'
+export { default as DoubleDatePickerWithoutError } from './atoms/form/DoubleDatePickerWithoutError'
 export { default as DoubleDatePicker } from './atoms/form/DoubleDatePicker'
-export { default as DoubleDatePickerError } from './atoms/form/DoubleDatePickerError'
 export { default as DropDown } from './atoms/form/DropDown'
 export { default as DropDownError } from './atoms/form/DropDownError'
 export { default as InfoTooltip } from './atoms/form/InfoTooltip'
